Remove overview page and add the code to page.tsx of landlords
