+++ components/Landlords/LandlordSidebar.tsx
@@ -11,7 +11,7 @@ const LanlordSidebar = () => {
         <aside id="default-sidebar" className={`w-72 h-screen transition-transform ${sidebarOpen ? '' : '-translate-x-full lg:translate-x-0'} lg:block lg:w-72 fixed left-0 z-50 bg-white h-screen border-r border-r-gray-300`} aria-label="Sidebar">
             <div className="h-full px-3 py-4 overflow-y-auto bg-gray-50">
                 <ul className="space-y-5 font-medium">
-                    <li><Link href='/landlords/overview'>
+                    <li><Link href='/landlords'>
                         <div className="flex items-center p-2 text-gray-900 rounded-lg hover:bg-gray-100 group">
                             <svg className="w-5 h-5 text-gray-500 transition duration-75 group-hover:text-gray-900 " aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 22 21">
                                 <path d="M16.975 11H10V4.025a1 1 0 0 0-1.066-.998 8.5 8.5 0 1 0 9.039 9.039.999.999 0 0 0-1-1.066h.002Z" />
